moving global staleness to a regular service, not angular
